+++ SPOUTSDK/SpoutGL/SpoutSender.h
@@ -90,7 +90,7 @@ class SPOUT_DLLEXP SpoutSender {
 	// Return frame count status
 	bool IsFrameCountEnabled();
 	// Frame rate control
-	void HoldFps(int fps);
+	void HoldFps(double fps);
 	// Signal sync event 
 	void SetFrameSync(const char* SenderName);
 	// Wait or test for a sync event
